requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
authlib==1.3.1
python-dotenv==1.0.1
pymongo==4.7.2
reportlab==4.4.10
requests==2.31.0
